requirements.txt
agno==2.5.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
btrees==6.3
certifi==2026.1.4
cffi==2.0.0
chameleon==4.6.0
charset-normalizer==3.4.4
click==8.3.1
curl-cffi==0.13.0
distro==1.9.0
docstring-parser==0.17.0
frozendict==2.4.7
gitdb==4.0.12
gitpython==3.1.46
grok==6.2
grokcore-annotation==5.0
grokcore-catalog==5.0
grokcore-chameleon==5.0
grokcore-component==5.0
grokcore-content==5.0
grokcore-formlib==5.0
grokcore-layout==5.0
grokcore-message==5.0
grokcore-security==5.1
grokcore-site==5.0
grokcore-traverser==5.0
grokcore-view==5.0
grokcore-viewlet==5.1
groq==1.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
markdown-it-py==4.0.0
martian==2.1
mdurl==0.1.2
multipart==1.3.0
multitasking==0.0.12
numpy==2.4.2
packaging==26.0
pandas==3.0.1
peewee==3.19.0
persistent==6.5
pip==24.0
platformdirs==4.9.2
protobuf==6.33.5
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-gettext==5.0
python-multipart==0.0.22
pytz==2025.2
pyyaml==6.0.3
regex==2026.2.19
requests==2.32.5
rich==14.3.3
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
tavily-python==0.7.21
tiktoken==0.12.0
transaction==5.0
typer==0.24.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uv==0.10.4
websockets==16.0
yfinance==1.2.0
z3c-flashmessage==4.0
z3c-pt==5.1
zc-catalog==5.0
zc-lockfile==4.0
zconfig==4.3
zdaemon==5.2.1
zodb==6.2
zodbpickle==4.3
zope-annotation==6.0
zope-app-appsetup==6.0
zope-app-publication==6.0
zope-app-wsgi==6.0
zope-authentication==6.0
zope-browser==4.0
zope-browserpage==6.0
zope-browserresource==6.0
zope-cachedescriptors==6.0
zope-catalog==6.0
zope-component==7.1
zope-configuration==7.0
zope-container==7.2
zope-contentprovider==7.0
zope-contenttype==6.0
zope-datetime==6.0
zope-deferredimport==6.1.1
zope-deprecation==6.0
zope-dottedname==7.0
zope-error==6.0
zope-errorview==3.0
zope-event==6.1
zope-exceptions==6.0
zope-filerepresentation==7.0
zope-formlib==7.0
zope-generations==7.0
zope-hookable==8.2
zope-i18n==6.0
zope-i18nmessageid==8.2
zope-index==8.1
zope-interface==8.2
zope-intid==6.0
zope-keyreference==7.0
zope-lifecycleevent==6.0
zope-location==6.0
zope-login==4.0
zope-minmax==4.0
zope-pagetemplate==6.1
zope-password==6.1
zope-principalregistry==6.0
zope-processlifetime==4.0
zope-proxy==7.1
zope-ptresource==6.0
zope-publisher==8.0
zope-schema==8.1
zope-security==8.3
zope-securitypolicy==6.0
zope-session==6.0
zope-site==6.0
zope-size==6.0
zope-tal==6.0
zope-tales==7.0
zope-testing==6.1
zope-traversing==6.0
zope-viewlet==6.0
